Move http server and graceful shutdown to index.js
* test watch was broken because http server began listening during each
test run
* update npm start script to use index.js
