+++ cyc_entry.js
@@ -1,5 +1,4 @@
 const express = require('express');
-const http = require('http');
 const path = require('path');
 const logger = require('morgan');
 const compression = require('compression');
@@ -44,30 +43,5 @@ if (app.get('env') === 'production') {
     res.sendStatus(err.status || 500);
   });
 }
-
-const server = http.createServer(app)
-server.listen(app.get('port'), () => {
-  console.log(`Express server listening on port ${app.get('port')}`);
-});
-
-const gracefulShutdown = function gracefulShutdown() {
-  console.log('Received kill signal, shutting down gracefully.');
-  server.close(function() {
-    console.log('Closed out remaining connections.');
-    process.exit();
-  });
-
-  // if after
-  setTimeout(function() {
-    console.error('Could not close connections in time, forcefully shutting down');
-    process.exit();
-  }, 10 * 1000);
-};
-
-// listen for TERM signal .e.g. kill
-process.on('SIGTERM', gracefulShutdown);
-
-// listen for INT signal e.g. Ctrl-C
-process.on('SIGINT', gracefulShutdown);
 
 module.exports = app;
